# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk Sistem Pelaporan Prestasi Mahasiswa",
        "title": "Student Achievement Reporting API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/achievements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Get achievements (role-based)",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Create new achievement draft",
                "responses": {}
            }
        },
        "/achievements/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Get achievement detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mongo Achievement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Update draft achievement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mongo Achievement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Delete achievement draft",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mongo Achievement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/achievements/{id}/attachments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Upload achievement attachment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Upload file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/achievements/{id}/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Get achievement history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/achievements/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Reject student achievement",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/achievements/{id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Submit achievement for verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/achievements/{id}/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Achievements"
                ],
                "summary": "Verify student achievement (Lecturer)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}